Refactor file package and move functions to utils package
